+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Format the `active` attribute as `Status` on the find-user modal. Fixes UICHKOUT-472.
 * Update due date on checkout after changed. Fixes UICHKOUT-470.
 * Fix hardcoded strings. Fixes UICHKOUT-458.
-
+* Simpler input focus due to better ref handling. Refs STCOM-394.
 
 ## [1.3.1](https://github.com/folio-org/ui-checkout/tree/v1.3.1) (2018-10-09)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v1.3.0...v1.3.1)